% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{age_spline_gamma}
\alias{age_spline_gamma}
\title{age_spline_gamma}
\usage{
age_spline_gamma(age, pars)
}
\arguments{
\item{age}{a vector of ages}

\item{pars}{a vector of length three giving the gamma coefficients
mode, shape, peak (i.e. height of peak at the mode), mode should be > 0,
shape > 1 and peak should be in unit interval.}
}
\value{
gamma-based spline at input ages
}
\description{
age_spline_gamma
}
