% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{age_spline_lognormal}
\alias{age_spline_lognormal}
\title{age_spline_lognormal}
\usage{
age_spline_lognormal(age, pars)
}
\arguments{
\item{age}{a vector of ages}

\item{pars}{a vector of length three giving the lognormal coefficients
mode, sd, peak (i.e. height of peak at the mode), mode and sd should be > 0,
peak should be in unit interval.}
}
\value{
lognormal-based spline at input ages
}
\description{
age_spline_lognormal
}
