% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{age_spline_polynomial}
\alias{age_spline_polynomial}
\title{age_spline_polynomial}
\usage{
age_spline_polynomial(age, pars)
}
\arguments{
\item{age}{a vector of ages}

\item{pars}{a vector of length three giving the polynomial coefficients}
}
\value{
polynomial spline at input ages
}
\description{
age_spline_polynomial
}
