% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{check_gas_parameters}
\alias{check_gas_parameters}
\title{Check model parameters are within permitted ranges}
\usage{
check_gas_parameters(pars)
}
\arguments{
\item{pars}{a named list of model parameters}
}
\description{
Check model parameters are within permitted ranges
}
