% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{demographic_parameters}
\alias{demographic_parameters}
\title{Demographic model parameters}
\usage{
demographic_parameters(n_group = 1)
}
\arguments{
\item{n_group}{integer giving number of age groups to divide pop}
}
\value{
A list of demographic parameters
}
\description{
Demographic model parameters
}
