% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{example_gas_parameters}
\alias{example_gas_parameters}
\title{Example of fitted gas parameters for use in testing}
\usage{
example_gas_parameters()
}
\value{
A list of named model parameters
}
\description{
Example of fitted gas parameters for use in testing
}
