% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{example_parameters}
\alias{example_parameters}
\title{Example of fitted gas parameters for use in testing}
\usage{
example_parameters(n_group = 1)
}
\arguments{
\item{n_group}{the number of model age-groups, defaults to 1}
}
\value{
A list of named model parameters
}
\description{
Example of fitted gas parameters for use in testing
}
