% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helium.R
\name{helium_compare}
\alias{helium_compare}
\title{Compare observed and modelled data for the helium model}
\usage{
helium_compare(state, observed, pars)
}
\arguments{
\item{state}{State vector for the end of the current week. This is
assumed to be filtered following \code{\link[=helium_index]{helium_index()}} so contains
rows corresponding to average daily pharyngitis rate per 100,000,
weekly number of scarlet fever cases and weekly number of iGAS cases.}

\item{observed}{Observed data. This will be a list with elements:
\code{scarlet_fever_inc}, \code{igas_inc}, \code{daily_pharyngitis_rate},
\code{daily_pharyngitis_rate_04}, \code{daily_pharyngitis_rate_05_14},
\code{daily_pharyngitis_rate_15_44}, \code{daily_pharyngitis_rate_45_64},
\code{daily_pharyngitis_rate_65_74}, \code{daily_pharyngitis_rate_75},
\code{daily_scarlet_fever_rate_04}, \code{daily_scarlet_fever_rate_05_14},
\code{daily_scarlet_fever_rate_15_44}, \code{daily_scarlet_fever_rate_45_64},
\code{daily_scarlet_fever_rate_65_74}, \code{daily_scarlet_fever_rate_75}}

\item{pars}{A list of parameters, as created by \code{\link[=model_parameters]{model_parameters()}}}
}
\value{
A vector of log likelihoods, the same length as the number
of particles (the number of columns in the modelled state)
}
\description{
Compare observed and modelled data from the helium model. This
conforms to the mcstate interface.
}
