% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helium.R
\name{helium_filter}
\alias{helium_filter}
\title{Create particle filter for helium model}
\usage{
helium_filter(data, constant_data, n_particles, compiled_compare = FALSE)
}
\arguments{
\item{data}{The data set to be used for the particle filter,
This is essentially a \code{\link[=data.frame]{data.frame()}} with at least columns \code{model_week},
along with the data used in \code{\link[=helium_compare]{helium_compare()}}.}

\item{constant_data}{a list with entries \code{etiologic_fraction} and
\code{asymptomatic_carriage} each containing a data.frame with entries
\code{age_start}, \code{age_end}, \code{N} and \code{n}.}

\item{n_particles}{The number of particles to simulate}

\item{compiled_compare}{Use a compiled compare function for the filter?}
}
\value{
a particle filter for the helium model
}
\description{
Create particle filter for helium model
}
