% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helium.R
\name{helium_fitted_states}
\alias{helium_fitted_states}
\title{Helium fitted states}
\usage{
helium_fitted_states()
}
\value{
A character vector of data states to which the model is fitted
}
\description{
Helium fitted states
}
