% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helium.R
\name{helium_index}
\alias{helium_index}
\title{Helium fitted states}
\usage{
helium_index(info)
}
\value{
A list with elements \code{run}, indicating the locations of the
compartments used in fitting the model, and \code{state} indicating the locations
of additional interesting model outputs
}
\description{
Helium fitted states
}
\examples{
p <- example_parameters(19)
mod <- model$new(p, 0, 10)
helium_index(mod$info())
}
