% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helium.R
\name{helium_prepare_data}
\alias{helium_prepare_data}
\title{Prepare particle filter data for the helium model}
\usage{
helium_prepare_data(data)
}
\arguments{
\item{data}{The data set to be used for the particle filter,
This is essentially a \code{\link[=data.frame]{data.frame()}} with at least columns \code{model_week},
along with the data used in \code{\link[=helium_compare]{helium_compare()}}.}
}
\value{
A a \code{\link[=data.frame]{data.frame()}} with columns \code{time_start}, \code{time_end},
along with the data used in \code{\link[=helium_compare]{helium_compare()}}.
}
\description{
Prepare particle filter data for the helium model
}
