% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrogen.R
\name{hydrogen_compare}
\alias{hydrogen_compare}
\title{Compare observed and modelled data for the hydrogen model}
\usage{
hydrogen_compare(state, observed, pars)
}
\arguments{
\item{state}{State vector for the end of the current week. This is
assumed to be filtered following \code{\link[=hydrogen_index]{hydrogen_index()}} so contains
rows corresponding to average daily pharyngitis rate per 100,000,
weekly number of scarlet fever cases and weekly number of iGAS cases.}

\item{observed}{Observed data. This will be a list with elements
\code{daily_pharyngitis_rate}, \code{daily_scarlet_fever_rate}, \code{scarlet_fever_cases}
and \code{igas_inc}}

\item{pars}{A list of parameters, as created by \code{\link[=model_parameters]{model_parameters()}}}
}
\value{
A vector of log likelihoods, the same length as the number
of particles (the number of columns in the modelled state)
}
\description{
Compare observed and modelled data from the hydrogen model. This
conforms to the mcstate interface.
}
\examples{
state <- rbind(daily_gas_pharyngitis_rate = 10:15,
               daily_scarlet_fever_rate = 0:5,
               scarlet_fever_cases = 100:105,
               igas_inc = 90:95)
observed <- list(daily_pharyngitis_rate = 13,
                 daily_scarlet_fever_rate = 3,
                 scarlet_fever_cases = 103,
                 igas_inc = 93)
pars <- example_parameters(1)
hydrogen_compare(state, observed, pars)
hydrogen_compare(state * 5, observed, pars)
}
