% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrogen.R
\name{hydrogen_create_transform}
\alias{hydrogen_create_transform}
\title{Create transform function for hydrogen model}
\usage{
hydrogen_create_transform(demographic_pars)
}
\arguments{
\item{demographic_pars}{A list of demographic parameters containing elements
\code{N0}, \code{alpha}, \code{omega}, \code{m}, \code{r_age} to be loaded into the transform}
}
\value{
A transform function for use in \code{\link[mcstate:pmcmc]{mcstate::pmcmc()}}
}
\description{
Create transform function for hydrogen model
}
