% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrogen.R
\name{hydrogen_filter}
\alias{hydrogen_filter}
\title{Create particle filter for hydrogen model}
\usage{
hydrogen_filter(data, n_particles)
}
\arguments{
\item{data}{The data set to be used for the particle filter,
This is essentially a \code{\link[=data.frame]{data.frame()}} with at least columns \code{model_week},
along with the data used in \code{\link[=hydrogen_compare]{hydrogen_compare()}}.}

\item{n_particles}{The number of particles to simulate}
}
\value{
a particle filter for the hydrogen model
}
\description{
Create particle filter for hydrogen model
}
