% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrogen.R
\name{hydrogen_fitted_states}
\alias{hydrogen_fitted_states}
\title{Hydrogen fitted states}
\usage{
hydrogen_fitted_states()
}
\value{
A character vector of the model states that are fitted to data
}
\description{
Hydrogen fitted states
}
