% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrogen.R
\name{hydrogen_index}
\alias{hydrogen_index}
\title{Hydrogen fitted states}
\usage{
hydrogen_index(info)
}
\arguments{
\item{info}{The result of running the \verb{$info()} method on an
initialised model}
}
\value{
A list with elements \code{run}, indicating the locations of the
compartments used in fitting the model, and \code{state} indicating the locations
of additional interesting model outputs
}
\description{
Hydrogen fitted states
}
\examples{
p <- example_gas_parameters(1)
mod <- model$new(p, 0, 10)
hydrogen_index(mod$info())
}
