% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrogen.R
\name{hydrogen_prepare_data}
\alias{hydrogen_prepare_data}
\title{Prepare particle filter data for the hydrogen model}
\usage{
hydrogen_prepare_data(data)
}
\arguments{
\item{data}{The data set to be used for the particle filter,
This is essentially a \code{\link[=data.frame]{data.frame()}} with at least columns \code{model_week},
along with the data used in \code{\link[=hydrogen_compare]{hydrogen_compare()}}.}
}
\value{
A a \code{\link[=data.frame]{data.frame()}} with columns \code{step_start}, \code{step_end},
along with the data used in \code{\link[=hydrogen_compare]{hydrogen_compare()}}.
}
\description{
Prepare particle filter data for the hydrogen model
}
