% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{index}
\alias{index}
\title{Index of the model}
\usage{
index(info)
}
\arguments{
\item{info}{The result of running the \verb{$info()} method on an initialised
model}
}
\value{
A list with element \code{run}, indicating the locations of (in order)
(1) pharyngitis_scarlet_fever_rate, (2) scarlet_fever_inc,
(3) iGAS_inc and with element \code{state} containing the same values,
followed by U, R, entrants_inc, leavers_inc, infections_inc and
pharyngitis_inc.
}
\description{
Index of interesting elements for the model. This function
conforms to the mcstate interface.
}
