% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{initial_parameters}
\alias{initial_parameters}
\title{Initial conditions for the model}
\usage{
initial_parameters(pars)
}
\arguments{
\item{pars}{A parameter list containing \code{N0}, \code{prev_A} and \code{prev_R} elements.}
}
\value{
A named list of initial model states
}
\description{
Initial conditions for the model
}
