% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{ll_dirichlet}
\alias{ll_dirichlet}
\title{ll_dirichlet}
\usage{
ll_dirichlet(data, state, exp_noise)
}
\arguments{
\item{data}{a vector containing observations from a Dirichlet distribution
i.e. a vector of probabilities that sum to 1}

\item{state}{a matrix containing sets of shape parameters, with one row
per parameter set}

\item{exp_noise}{exponential noise for case when all probabilities are 0.}
}
\description{
ll_dirichlet
}
