% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{ll_multinom}
\alias{ll_multinom}
\title{ll_multinom}
\usage{
ll_multinom(data, prob, noise)
}
\arguments{
\item{data}{a vector containing observations from a multinomial distribution}

\item{prob}{a matrix containing sets of probabilities for the multinomial
distribution, with one set per column. Values will scale automatically to
so that each column sums to 1. the number of rows should be the same length
as the data.}

\item{noise}{noise for case when all probabilities are 0.}
}
\description{
ll_multinom
}
