% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{mean_age_spline}
\alias{mean_age_spline}
\title{mean_age_spline}
\usage{
mean_age_spline(age_start, age_end, pars, spline)
}
\arguments{
\item{age_start}{a vector denoting the start of each age bracket}

\item{age_end}{a vector denoting the end of each age bracket}

\item{pars}{a vector of length three giving the coefficients for \code{spline}}

\item{spline}{a spline function, options include \code{age_spline_polynomial},
\code{age_spline_gamma} or \code{age_spline_lognormal}}
}
\value{
vector giving mean of spline output for ages within each bracket
}
\description{
mean_age_spline
}
