% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{model_date}
\alias{model_date}
\title{Convert model_day to date}
\usage{
model_date(day)
}
\arguments{
\item{day}{A vector of model_days (i.e. integers),
where model_day 1 = 30 Dec 2013, to align with the PHE/UKHSA definition of
epi weeks, whereby week 1 of 2014 ends on Sunday 5 Jan 2014}
}
\description{
Convert model_day to date
}
\examples{
model_date(1)
model_date(7)
}
