% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{model_day}
\alias{model_day}
\title{Convert date to model_day}
\usage{
model_day(date)
}
\arguments{
\item{date}{A vector of dates or character strings containing dates}
}
\value{
A vector of model_days (i.e. integers),
where model_day 1 = 30 Dec 2013, to align with the PHE/UKHSA definition of
epi weeks, whereby week 1 of 2014 ends on Sunday 5 Jan 2014
}
\description{
Convert date to model_day
}
\examples{
model_day("2013-12-30")
model_day("2014-01-05")
}
