% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{model_index}
\alias{model_index}
\title{Named list of model output indices}
\usage{
model_index(n_group = 1)
}
\arguments{
\item{n_group}{Integer number of age groups}
}
\value{
Named list of model output indices
}
\description{
Named list of model output indices
}
