% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{model_parameters}
\alias{model_parameters}
\title{Demographic model parameters}
\usage{
model_parameters(gas_pars, initial_pars = NULL, demographic_pars = NULL)
}
\arguments{
\item{gas_pars}{A list of named fitted model parameters}

\item{initial_pars}{default NULL uses \code{N0}, \code{prev_A} and \code{prev_R} to
calculate the starting position}

\item{demographic_pars}{default NULL uses \code{demographic_parameters()})}
}
\value{
A list of all model parameters
}
\description{
Demographic model parameters
}
