% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{model_week}
\alias{model_week}
\title{Convert date to model_week}
\usage{
model_week(date)
}
\arguments{
\item{date}{A vector of dates or character strings containing dates}
}
\value{
A vector of model_weeks (i.e. integers),
where model_week 1 = ends on Sunday 5 Jan 2014
to align with the PHE/UKHSA definition of epi weeks
}
\description{
Convert date to model_week
}
\examples{
model_week("2014-01-01")
model_week("2014-01-07")
}
