% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{no_gas_parameters}
\alias{no_gas_parameters}
\title{Model parameters with no gas for use in testing}
\usage{
no_gas_parameters(n_group = 1)
}
\arguments{
\item{n_group}{the number of model age-groups, defaults to 1}
}
\value{
A list of named model parameters
}
\description{
Model parameters with no gas for use in testing
}
