% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{transform}
\alias{transform}
\title{Transform fitted parameters into gas params}
\usage{
transform(pars)
}
\arguments{
\item{pars}{list of fitted parameters}
}
\value{
A list of parameters for use in the model
}
\description{
Transform fitted parameters into gas params
}
\details{
Transform fitted parameters into gas params
}
