% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcmc.R
\name{compare}
\alias{compare}
\title{Calculate the log likelihood of the data given the parameters}
\usage{
compare(state, observed, pars)
}
\arguments{
\item{state}{State vector for the end of the current day. This is
assumed to be filtered following \code{\link[=index]{index()}} so contains 3 rows corresponding
to daily_pharyngitis_scarlet_fever_rate, scarlet_fever and igas flows}

\item{observed}{Observed data containing entries
\code{daily_pharyngitis_scarlet_fever_rate}, \code{scarlet_fever_inc} and \code{igas_inc}.}

\item{pars}{A list of parameters, as created by \code{\link[=model_parameters]{model_parameters()}}}
}
\value{
a single log likelihood
}
\description{
Calculate the log likelihood of the data given the parameters
}
