% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{demographic_parameters}
\alias{demographic_parameters}
\title{Demographic model parameters}
\usage{
demographic_parameters()
}
\value{
A list of demographic parameters
}
\description{
Demographic model parameters
}
