% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{model_date}
\alias{model_date}
\title{Convert model_day to date}
\usage{
model_date(day)
}
\arguments{
\item{date}{A vector of model_days (i.e. integers),
where model_day 1 = 1 Jan 2014}
}
\value{
A vector of dates
}
\description{
Convert model_day to date
}
\examples{
model_date(1)
model_date(7)
}
