% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{model_day}
\alias{model_day}
\title{Convert date to model_day}
\usage{
model_day(date)
}
\arguments{
\item{date}{A vector of dates or character strings containing dates}
}
\value{
A vector of model_days (i.e. integers),
where model_day 1 = 1 Jan 2014
}
\description{
Convert date to model_day
}
\examples{
model_day("2014-01-01")
model_day("2014-01-07")
}
