% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{model_week_date}
\alias{model_week_date}
\title{Convert model_week to date on which the week ends}
\usage{
model_week_date(week)
}
\arguments{
\item{week}{A vector of model_weeks (i.e. integers),
where model_week 1 = ends on Sunday 5 Jan 2014
to align with the PHE/UKHSA definition of epi weeks}
}
\value{
A vector of dates giving the date on which the week ends
}
\description{
Convert model_week to date on which the week ends
}
\examples{
model_week_date(1)
model_week_date(7)
}
